+++ src/DBmysqlIterator.php
@@ -345,7 +345,9 @@ class DBmysqlIterator implements SeekableIterator, Countable
 
         $cleanorderby = [];
         foreach ($clause as $o) {
-            if (is_string($o)) {
+            if (is_string($o) && str_contains($o, ',')) {
+                Toolbox::deprecated('Using comma separated list of fields is deprecated in ORDER BY clause.');
+
                 $fields = explode(',', $o);
                 foreach ($fields as $field) {
                     $new = '';
@@ -357,6 +359,14 @@ class DBmysqlIterator implements SeekableIterator, Countable
                     }
                     $cleanorderby[] = $new;
                 }
+            } elseif (is_string($o)) {
+                $parts = explode(' ', trim($o));
+                $name = trim($parts[0]);
+                $direction = trim($parts[1] ?? '');
+
+                $cleanorderby[] = DBmysql::quoteName($name)
+                    . (in_array($direction, ['ASC', 'DESC']) ? ' ' . $direction : '');
+
             } elseif ($o instanceof QueryExpression) {
                 $cleanorderby[] = $o->getValue();
             } else {